+++ alan-adventure/VectorMath.cpp
@@ -6,14 +6,14 @@ const float vectorDistance(const float x1, const float y1, const float x2, const
   const float dx = x2 - x1;
   const float dy = y2 - y1;
 
-  return sqrt(pow(dx, 2) + pow(dy, 2));
+  return sqrtf(powf(dx, 2) + powf(dy, 2));
 }
 
 const float vectorDistance(sf::Vector2f vec1, sf::Vector2f vec2)
 {
   const sf::Vector2f dVec = vec2 - vec1;
 
-  return sqrt(pow(dVec.x, 2) + pow(dVec.y, 2));
+  return sqrtf(powf(dVec.x, 2) + powf(dVec.y, 2));
 }
 
 const float vectorAngle(sf::Vector2f vec1, sf::Vector2f vec2)
